Use explicit type conversion

+++ src/lib/base/Path.cpp
@@ -30,11 +30,11 @@ std::wstring path(const String& filePath)
 {
     std::wstring result;
 
-    auto lenght = MultiByteToWideChar(CP_UTF8, 0, filePath.c_str(), filePath.length(), NULL, 0);
+    auto lenght = MultiByteToWideChar(CP_UTF8, 0, filePath.c_str(), (int)filePath.length(), NULL, 0);
     if (lenght > 0)
     {
         result.resize(lenght);
-        MultiByteToWideChar(CP_UTF8, 0, filePath.c_str(), filePath.length(), &result[0], lenght);
+        MultiByteToWideChar(CP_UTF8, 0, filePath.c_str(), (int)filePath.length(), &result[0], lenght);
     }
 
     return result;
